+++ cookbook/examples/apps/tic_tac_toe/requirements.in
@@ -1,19 +1,14 @@
-# Core dependencies
-streamlit
-nest-asyncio
-Pillow
-
 agno
-openai
 anthropic
-ollama
-google-genai
-
-python-dotenv
-pathlib
-typing
 dataclasses
+google-genai
+nest-asyncio
+ollama
+openai
+pathlib
+Pillow
+pip-tools
+python-dotenv
 rich
-
-uv
-pip-tools 
+streamlit
+typing
