+++ cookbook/examples/apps/tic_tac_toe/requirements.txt
@@ -1,6 +1,6 @@
 # This file was autogenerated by uv via the following command:
 #    ./generate_requirements.sh
-agno==1.1.0
+agno==1.1.1
     # via -r cookbook/examples/apps/tic_tac_toe/requirements.in
 altair==5.5.0
     # via streamlit
@@ -53,7 +53,7 @@ gitpython==3.1.44
     #   streamlit
 google-auth==2.38.0
     # via google-genai
-google-genai==1.1.0
+google-genai==1.2.0
     # via -r cookbook/examples/apps/tic_tac_toe/requirements.in
 h11==0.14.0
     # via httpcore
@@ -92,14 +92,14 @@ narwhals==1.26.0
     # via altair
 nest-asyncio==1.6.0
     # via -r cookbook/examples/apps/tic_tac_toe/requirements.in
-numpy==2.2.2
+numpy==2.2.3
     # via
     #   pandas
     #   pydeck
     #   streamlit
 ollama==0.4.7
     # via -r cookbook/examples/apps/tic_tac_toe/requirements.in
-openai==1.61.1
+openai==1.63.0
     # via -r cookbook/examples/apps/tic_tac_toe/requirements.in
 packaging==24.2
     # via
@@ -216,6 +216,7 @@ typing-extensions==4.12.2
     #   altair
     #   anthropic
     #   anyio
+    #   google-genai
     #   openai
     #   pydantic
     #   pydantic-core
@@ -226,8 +227,6 @@ tzdata==2025.1
     # via pandas
 urllib3==2.3.0
     # via requests
-uv==0.5.30
-    # via -r cookbook/examples/apps/tic_tac_toe/requirements.in
 websockets==14.2
     # via google-genai
 wheel==0.45.1
